Features added. Current State: Client Only program complete
The program is currently completed. As requirements change and things are added I will add to the first program.

programplan.txt has been updated with implemented features, possible issues and current status of each module. Document version also updated.

prerequisites.txt added. A document specifically to list all the programs or libraries which is needed for this group of scripts/programs.