Partial commit with the server side being able to send requests.
I have updated the ServerStatusServer.py file to be able to open a socket and receive data (currently commented out due to testing and further advancements) and the ServerStatusClient can send certain data values to the server if it is hardcoded.

The reasoning for the ServerStatusServer.py having the open connection coded out is for testing. I am at a stage where I have started to build the GUI. At this stage it only has 12 buttons but I can interact with the buttons without the main window closing and it attempts to send data which is refused since the recieve function is not implemented on the client side of the program.

This is a partial commit as a backup and update log for future reference if I need to look back at any commits in the future. Multithreading has been implemented for the buttons, threads are created to send data and then closed when no longer needed to save resources and they only stay open for a very short duration of time.

The buttons in the GUI at this moment in time will only send the numbers "1-12" depending on what button is pressed. This is currently a placeholder so I can integrate the recieving side of the client program. When any of the buttons are pressed I want the client program to send the bound metric to the server program. This will be the baseline for the main gui which will be built. it will only output the metrics to the console until further development on the server GUI has been completed.